% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmar-package.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error (MAE)}
\usage{
mae(known, predicted)
}
\arguments{
\item{known}{a numeric vector}

\item{predicted}{a numeric vector}
}
\value{
MAE
}
\description{
Evaluates MAE on predictions of a regression model
}
\author{
David Senhora Navega
}
