% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmar-package.R
\name{medae}
\alias{medae}
\title{Median Absolute Error (MedAE)}
\usage{
medae(known, predicted)
}
\arguments{
\item{known}{a numeric vector}

\item{predicted}{a numeric vector}
}
\value{
MedAE
}
\description{
Evaluates MedAE on predictions of a regression model
}
\author{
David Senhora Navega
}
