% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmar-plot.R
\name{rma_accuracy_plot}
\alias{rma_accuracy_plot}
\title{Regression Accuracy Plot}
\usage{
rma_accuracy_plot(
  known,
  predicted,
  interval = NULL,
  label = NULL,
  breaks = 10,
  base_size = 12,
  digits = 3
)
}
\arguments{
\item{known}{a numeric vector}

\item{predicted}{a numeric vector or matrix. See Details.}

\item{interval}{minimum and maximum value of prediction domain. Should be
supplied as vector of with two elements. If not supplied is assess from the
known argument.}

\item{label}{a character defining an alternative label for the x axis.}

\item{breaks}{breaks on plot axis. Default is 10.}

\item{base_size}{ggplot parameter to control text size in plots.
Default is 12}

\item{digits}{number of digits controlling floats precision. Default is 3}
}
\value{
a ggplot2 graphical object
}
\description{
Regression Accuracy Plot
}
\details{
If predicted is a matrix organized it should be organized as follows:
3 columns where the first column stores the point estimate of the model
and the second and third columns store the lower and upper bounds of the
predictive interval if the algorithm is able to produce one.
}
\author{
David Senhora Navega
}
