% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmar-plot.R
\name{rma_efficiency_plot}
\alias{rma_efficiency_plot}
\title{Regression Efficiency Plot}
\usage{
rma_efficiency_plot(
  known,
  predicted,
  interval = NULL,
  breaks = 10,
  base_size = 12,
  digits = 3
)
}
\arguments{
\item{known}{a numeric vector}

\item{predicted}{a numeric matrix. See Details.}

\item{interval}{minimum and maximum value of prediction domain. Should be
supplied as vector of with two elements. If not supplied is assess from the
known argument.}

\item{breaks}{breaks on plot axis. Default is 10.}

\item{base_size}{ggplot parameter to control text size in plots.
Default is 12}

\item{digits}{number of digits controlling floats precision. Default is 3}
}
\value{
a ggplot2 graphical object
}
\description{
In regression, efficiency can be evaluated through the analysis of the width
of the predictive interval. The function evaluates and plot the width of the
predictive interval and its coverage, the probability of an estimated
predictive interval contain the known value.
}
\details{
predicted matrix must be organized as follows:
3 columns where the first column stores the point estimate of the model
and the second and third columns store the lower and upper bounds of the
predictive interval if the algorithm is able to produce one.
}
\author{
David Senhora Navega
}
