% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmar.R
\name{rmar}
\alias{rmar}
\title{rmar: Regression Model Assessment in R}
\usage{
rmar(known, predicted, interval = NULL, digits = 3)
}
\arguments{
\item{known}{a numeric vector}

\item{predicted}{a numeric vector or matrix. See Details.}

\item{interval}{minimum and maximum value of prediction domain. Should be
supplied as vector of with two elements. If not supplied is assess from the
known argument.}

\item{digits}{an integer defining the precision of the round function.}
}
\description{
rmar: Regression Model Assessment in R
}
\details{
If predicted is a matrix, it should be organized in way that the first column
represents the point estimate and the second and third columns represent the
lower and upper bound of the predictive interval.
}
\author{
David Senhora Navega
}
