% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmar-package.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error (RMSE)}
\usage{
rmse(known, predicted)
}
\arguments{
\item{known}{a numeric vector}

\item{predicted}{a numeric vector}
}
\value{
RMSE
}
\description{
Evaluates RMSE on predictions of a regression model
}
\author{
David Senhora Navega
}
