% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmar-package.R
\name{rsquared}
\alias{rsquared}
\title{R Squared}
\usage{
rsquared(known, predicted)
}
\arguments{
\item{known}{a numeric vector}

\item{predicted}{a numeric vector}
}
\value{
R Squared value
}
\description{
Evaluates R Squared (Explained Variance)
}
\references{
Gelman A, Goodrich B, Gabry J, Vehtari A. R-squared for Bayesian regression
models. Am Stat 2019;73(3):307–9.
}
\author{
David Senhora Navega
}
